+++ service/src/main/java/eu/europa/ec/fisheries/uvms/rules/service/mapper/CustomMapper.java
@@ -358,7 +358,6 @@ public class CustomMapper {
         List<CodeType> codeTypes = new ArrayList<>();
         for (FACatch faCatches : faCatch) {
 
-
             switch (methodToChoose) {
 
                 case ActivityFactMapper.CODE_TYPE_FOR_FACATCH:
